pyproject.toml
[project]
name = "document-semantic"
version = "0.1.0"
description = "A pluggable document semantic processing pipeline"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.0",
    "python-docx>=1.1.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0",
    "markdown-it-py>=3.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "linkify-it-py>=2.1.0",
    "pydantic-settings>=2.14.0",
    "typer>=0.24.1",
    "rich>=15.0.0",
    "strands-agents[openai]>=1.36.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel",
    "pillow>=12.2.0",
    "pytest>=8.0",
    "pytest-snapshot>=0.9.0",
]
pandoc = [
    # pandoc is an external binary, no python package needed
]

[project.scripts]
document-semantic = "document_semantic.cli:app"
ds = "document_semantic.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/document_semantic"]

[[tool.uv.index]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true


[tool.ruff]
line-length = 120
target-version = "py312"  # Adjust based on your project's minimum Python version

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
]
ignore = ["E501"]  # Line length handled by formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"


